configuration menu

+++ plc_1xx_L433/Src/main.c
@@ -191,9 +191,9 @@ uint32_t boot_timer_counter;
 extern uint8_t bootloader_state;
 extern uint8_t receiveBuffer[16];
 
-extern uint8_t channel_ICP_ON;
-extern uint8_t channel_4_20_ON;
-extern uint8_t channel_485_ON;
+extern uint16_t channel_ICP_ON;
+extern uint16_t channel_4_20_ON;
+extern uint16_t channel_485_ON;
 
 volatile uint16_t timer_485_counter = 0;
 extern uint8_t break_sensor_485;
